+++ CPE200Lab1/CPE200Lab1/Program.cs
@@ -16,7 +16,7 @@ namespace CPE200Lab1
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new ExtendForm());
+            Application.Run(new MainForm());
         }
     }
 }
